Rename to PersistentHideInterface.js & fix
FighterNoForce isn't very direct, I believe PersistentHideInterface is more of that on the other hand.

The script did not seem to work originally.

Looking at it on debug mode, kanan detected there are two similar values. Applying the patch to the second one makes the script work as intended. However disabled the patch to the first one breaks it again. As a result, I came to the conclusion each of them is for the second and third chain combo respectively, and they both need to be patched together.